+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.1.1] - 2023-12-02
 ### Added
 - Year in Review page for indiebookclub as a whole, e.g. `/review/2023`
   - Page is set to start being available November 30 each year.